Correção de bugs deletando imagem

+++ lib/app/src/data/section_data.dart
@@ -90,12 +90,14 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items) {
-      try {
-        final ref = await storage.getReferenceFromUrl(item.image as String);
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.getReferenceFromUrl(item.image as String);
 
-        await ref.delete();
-        // ignore: empty_catches
-      } catch (e) {}
+          await ref.delete();
+          // ignore: empty_catches
+        } catch (e) {}
+      }
     }
   }
